moved some util functions to avoid circular import
