+++ tests/test_organization.py
@@ -143,3 +143,228 @@ class TestOrganization(BaseTest, TestCase):
         assert membership is not None
         assert membership.is_owner == False
         assert membership.joined
+
+class MembershipToDictTest(TestCase):
+
+    def create_app(self):
+        app.config.from_object('project.config.TestingConfig')
+        return app
+
+    def test_nonetype(self):
+        """
+        tests converting a NoneType object
+        :return:
+        """
+        assert org_utils.membership_to_dict(None) is None
+
+    def test_invalid_type(self):
+        """
+        Tets converting an object that isn't a membership
+        :return:
+        """
+        shift = Shift(assigned_user_id=1, description='', position_id=1,
+                             start_time='2007-04-05T14:30', end_time='2007-04-05T14:30')
+
+        self.assertRaises(TypeError, org_utils.membership_to_dict, shift)
+
+    def test_valid(self):
+        """
+        tests with a valid membership
+        :return:
+        """
+        membership = Membership(member_id=1, organization_id=1, is_admin=True, is_owner=False, joined=True)
+        assert membership is not None
+        membership_dict = org_utils.membership_to_dict(membership)
+
+        assert membership_dict is not None
+
+        assert 'id' in membership_dict
+        assert membership_dict['id'] == membership.id
+
+        assert 'member_id' in membership_dict
+        assert membership_dict['member_id'] == membership.member_id
+
+        assert 'organization_id' in membership_dict
+        assert membership_dict['organization_id'] == membership.organization_id
+
+        assert 'is_owner' in membership_dict
+        assert membership_dict['is_owner'] == False
+
+        assert 'is_admin' in membership_dict
+        assert membership_dict['is_admin'] == True
+
+        assert 'joined' in membership_dict
+        assert membership_dict['joined'] == True
+
+
+class ShiftToDictTest(TestCase):
+
+    def create_app(self):
+        app.config.from_object('project.config.TestingConfig')
+        return app
+
+    def test_nonetype(self):
+        """
+        tests converting a NoneType object
+        :return:
+        """
+        assert org_utils.shift_to_dict(None) is None
+
+    def test_invalid_type(self):
+        """
+        Tets converting an object that isn't a membership
+        :return:
+        """
+        membership = Membership(member_id=1, organization_id=1, is_admin=True, is_owner=False, joined=True)
+
+        self.assertRaises(TypeError, org_utils.shift_to_dict, membership)
+
+    def test_valid_none_missing(self):
+        """
+        Tests a valid dictionary where all fields are filled
+        :return:
+        """
+        position = Position(
+            title='Test Position',
+            organization_id=1
+        )
+
+        user = User(
+            email='member@organization.com',
+            first_name='John',
+            last_name='Doe',
+            password='password',
+            confirmed=True
+        )
+
+        shift = Shift(assigned_user_id=user.id, description='', position_id=position.id,
+                      start_time='2007-04-05T14:30', end_time='2007-04-05T14:30')
+
+        shift.Position = position
+        shift.user = user
+
+        assert shift is not None
+
+        shift_dict = org_utils.shift_to_dict(shift)
+        assert shift_dict is not None
+
+        assert 'id' in shift_dict
+        assert shift_dict['id'] == shift.id
+
+        assert 'position_id' in shift_dict
+        assert shift_dict['position_id'] == shift.Position.id
+
+        assert 'position_title' in shift_dict
+        assert shift_dict['position_title'] == shift.Position.title
+
+        assert 'start' in shift_dict
+        assert shift_dict['start'] == shift.start_time
+
+        assert 'end' in shift_dict
+        assert shift_dict['end'] == shift.end_time
+
+        assert 'description' in shift_dict
+        assert shift_dict['description'] == shift.description
+
+        assert 'assigned_member_id' in shift_dict
+        assert shift_dict['assigned_member_id'] == shift.assigned_user_id
+
+        assert 'assigned_member' in shift_dict
+        assert shift_dict['assigned_member'] == shift.user.first_name + ' ' + shift.user.last_name
+
+    def test_valid_description_missing(self):
+        """
+        Tests a valid dictionary where the description is missing
+        :return:
+        """
+        position = Position(
+            title='Test Position',
+            organization_id=1
+        )
+
+        user = User(
+            email='member@organization.com',
+            first_name='John',
+            last_name='Doe',
+            password='password',
+            confirmed=True
+        )
+
+        shift = Shift(assigned_user_id=user.id, description=None, position_id=position.id,
+                             start_time='2007-04-05T14:30', end_time='2007-04-05T14:30')
+
+        shift.Position = position
+        shift.user = user
+
+        assert shift is not None
+
+        shift_dict = org_utils.shift_to_dict(shift)
+        assert shift_dict is not None
+
+        assert 'id' in shift_dict
+        assert shift_dict['id'] == shift.id
+
+        assert 'position_id' in shift_dict
+        assert shift_dict['position_id'] == shift.Position.id
+
+        assert 'position_title' in shift_dict
+        assert shift_dict['position_title'] == shift.Position.title
+
+        assert 'start' in shift_dict
+        assert shift_dict['start'] == shift.start_time
+
+        assert 'end' in shift_dict
+        assert shift_dict['end'] == shift.end_time
+
+        assert 'description' in shift_dict
+        assert shift_dict['description'] == ''
+
+        assert 'assigned_member_id' in shift_dict
+        assert shift_dict['assigned_member_id'] == shift.assigned_user_id
+
+        assert 'assigned_member' in shift_dict
+        assert shift_dict['assigned_member'] == shift.user.first_name + ' ' + shift.user.last_name
+
+    def test_valid_user_missing(self):
+        """
+        Tests a valid dictionary where no user is assigned
+        :return:
+        """
+        position = Position(
+            title='Test Position',
+            organization_id=1
+        )
+
+        shift = Shift(assigned_user_id=None, description='A description', position_id=position.id,
+                             start_time='2007-04-05T14:30', end_time='2007-04-05T14:30')
+
+        shift.Position = position
+
+        assert shift is not None
+
+        shift_dict = org_utils.shift_to_dict(shift)
+        assert shift_dict is not None
+
+        assert 'id' in shift_dict
+        assert shift_dict['id'] == shift.id
+
+        assert 'position_id' in shift_dict
+        assert shift_dict['position_id'] == shift.Position.id
+
+        assert 'position_title' in shift_dict
+        assert shift_dict['position_title'] == shift.Position.title
+
+        assert 'start' in shift_dict
+        assert shift_dict['start'] == shift.start_time
+
+        assert 'end' in shift_dict
+        assert shift_dict['end'] == shift.end_time
+
+        assert 'description' in shift_dict
+        assert shift_dict['description'] == shift.description
+
+        assert 'assigned_member_id' in shift_dict
+        assert shift_dict['assigned_member_id'] == 0
+
+        assert 'assigned_member' in shift_dict
+        assert shift_dict['assigned_member'] == 'Unassigned'
